Add files via upload
These are the header and source files for the STL lab. This lab demonstrates the use of the stack by creating shapes with the STL and manipulating them in the stack and then "drawing" them to the output terminal in the order that they end up on the stack. This program uses standard template library to create the shapes. Some of the files were provided as templates that I filled in and others were written by me. 

Adam Claxton